split tropical_cyclone module
